feat(apollo,look&feel): inputdatetext in inputdate
