--- apps/look-and-feel-stories/src/InputDateText/InputDateText.mdx
@@ -1,33 +1,0 @@
-import { Canvas, Controls, Meta } from "@storybook/addon-docs";
-import * as InputDateTextStories from "./InputdateText.stories";
-
-<Meta of={InputDateTextStories} name="InputDateText" />
-
-# InputDateText
-
-To use the InputDateText import it like that:
-
-```tsx
-import { InputDateText } from "@axa-fr/design-system-apollo-react/lf";
-
-const MyComponent = () => (
-  <>
-    <InputDateText
-      value="John Doe"
-      label="Label"
-      placeholder="Placeholder"
-      helper="Informations complémentaires"
-      name="name"
-      id="nameid"
-      required
-      buttonLabel="my btn"
-    />
-  </>
-);
-```
-
-## Placeholder
-
-<Canvas of={InputDateTextStories.InputDateTextPlaygroundStory} />
-
-<Controls of={InputDateTextStories.InputDateTextPlaygroundStory} />
